Move tests into one file.
